refactor(vector-store-service): restructure vector store implementations
- Move LangChain-specific implementations from langchain/ to vectorstore/base/
- Update import paths in vector store modules to reference new base directory
- Remove unnecessary dependencies from package.json (neo4j-driver, @langchain/mongodb)
- Update package description to reflect RAG architecture support
- Streamline peer dependencies configuration

This refactoring improves code organization by consolidating vector store
base implementations and removes unused dependencies to reduce bundle size.
