+++ packages/vector-store-service/src/vectorstore/base/faiss.ts
